GBIO

Documentation for GBIO.

GBIO.isfuzzymatch

GBIO.levenshteindistance

GBIO.readJLD2

GBIO.readdelimited

GBIO.readdelimited

GBIO.readdelimited

GBIO.readvcf

GBIO.writeJLD2

GBIO.writedelimited

GBIO.writedelimited

GBIO.writedelimited

GBIO.writevcf

GBIO.isfuzzymatch — Method

isfuzzymatch(a::String, b::String; threshold::Float64=0.3)::Bool

Fuzzy string matching using the Levenshtein distance, and a threshold as a fraction of the smaller string.

Examples

julia> isfuzzymatch("populations", "populations")
true

julia> isfuzzymatch("populations", "poplation")
true

julia> isfuzzymatch("populations", "entry")
false

GBIO.levenshteindistance — Method

levenshteindistance(a::String, b::String)::Int64

Calculate the Levenshtein distance between 2 strings. TODO: optimise as we only need to compute 2 rows at a time and do not need the full matrix

Examples

julia> levenshteindistance("populations", "populations")
0

julia> levenshteindistance("populations", "poplation")
2

julia> levenshteindistance("populations", "entry")
3

GBIO.readJLD2 — Method

readJLD2(type::Type, fname::String = missing)::Type

Load a core (Genomes, Phenomes, and Trials), simulation (SimulatedEffects), and model (TEBV) struct from a JLD2 file.

Examples

julia> genomes = GBCore.simulategenomes(n=2, sparsity=0.01, verbose=false);

julia> fname = writeJLD2(genomes);

julia> readJLD2(Genomes, fname) == genomes
true

julia> phenomes = Phenomes(n=2, t=2); phenomes.entries = ["entry_1", "entry_2"]; phenomes.traits = ["trait_1", "trait_2"];

julia> fname = writeJLD2(phenomes);

julia> readJLD2(Phenomes, fname) == phenomes
true

julia> trials = Trials(n=1, t=2); trials.entries = ["entry_1"];

julia> fname = writeJLD2(trials);

julia> readJLD2(Trials, fname) == trials
true

GBIO.readdelimited — Method

readdelimited(type::Type{Genomes}; fname::String, sep::String = "\t")::Genomes

Load a Genomes struct from a string-delimited (default=\t) file. Each row corresponds to a locus-allele combination. The first 4 columns correspond to the chromosome, position, all alleles in the locus (delimited by |), and the specific allele. The subsequency columns refer to the samples, pools, entries or genotypes.

Examples

julia> genomes = GBCore.simulategenomes(n=10, sparsity=0.1, verbose=false);

julia> fname = writedelimited(genomes);

julia> genomes_reloaded = readdelimited(Genomes, fname=fname);

julia> genomes == genomes_reloaded
true

GBIO.readdelimited — Method

readdelimited(type::Type{Phenomes}; fname::String, sep::String = "\t")::Phenomes

Load a Phenomes struct from a string-delimited (default=\t) file. Each row corresponds to a locus-allele combination. The first 4 columns correspond to the chromosome, position, all alleles in the locus (delimited by |), and the specific allele. The subsequency columns refer to the samples, pools, entries or genotypes.

Examples

julia> phenomes = Phenomes(n=10, t=3); phenomes.entries = string.("entry_", 1:10); phenomes.populations .= "pop1"; phenomes.traits = ["A", "B", "C"]; phenomes.phenotypes = rand(10,3); phenomes.phenotypes[1,1] = missing; phenomes.mask .= true;

julia> fname = writedelimited(phenomes);

julia> phenomes_reloaded = readdelimited(Phenomes, fname=fname);

julia> phenomes == phenomes_reloaded
true

GBIO.readdelimited — Method

readdelimited(type::Type{Trials}; fname::String, sep::String = "\t")::Trials

Load a Trials struct from a string-delimited (default=\t) file. Each row corresponds to a locus-allele combination. The first 4 columns correspond to the chromosome, position, all alleles in the locus (delimited by |), and the specific allele. The subsequency columns refer to the samples, pools, entries or genotypes.

Examples

julia> genomes = GBCore.simulategenomes(n=10, verbose=false);

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, sparsity=0.1, verbose=false);

julia> fname = writedelimited(trials);

julia> trials_reloaded = readdelimited(Trials, fname=fname);

julia> trials == trials_reloaded
true

GBIO.readvcf — Method

readvcf(;fname::String, field::String = "any", verbose::Bool = false)::Genomes

Load Genomes struct from vcf file

Examples

julia> genomes = GBCore.simulategenomes(n=10, sparsity=0.1, verbose=false);

julia> fname = writevcf(genomes);

julia> genomes_reloaded = readvcf(fname=fname);

julia> genomes.entries == genomes_reloaded.entries
true

julia> dimensions(genomes) == dimensions(genomes_reloaded)
true

julia> ismissing.(genomes.allele_frequencies) == ismissing.(genomes_reloaded.allele_frequencies)
true

GBIO.writeJLD2 — Method

writeJLD2(A::Union{Genomes,Phenomes,Trials,SimulatedEffects}; fname::Union{Missing,String} = missing)::String

Save core (Genomes, Phenomes, and Trials), simulation (SimulatedEffects), and model (TEBV) structs as JLD2, a heirarchical data format version 5 (HDF5) - compatible format. Note that the extension name should be '.jld2'.

Examples

julia> genomes = GBCore.simulategenomes(n=2, verbose=false);

julia> writeJLD2(genomes, fname="test_genomes.jld2")
"test_genomes.jld2"

julia> genomes_reloaded = load("test_genomes.jld2");

julia> genomes_reloaded[collect(keys(genomes_reloaded))[1]] == genomes
true

julia> phenomes = Phenomes(n=2, t=2); phenomes.entries = ["entry_1", "entry_2"]; phenomes.traits = ["trait_1", "trait_2"];

julia> writeJLD2(phenomes, fname="test_phenomes.jld2")
"test_phenomes.jld2"

julia> phenomes_reloaded = load("test_phenomes.jld2");

julia> phenomes_reloaded[collect(keys(phenomes_reloaded))[1]] == phenomes
true

julia> trials = Trials(n=1, t=2); trials.entries = ["entry_1"];

julia> writeJLD2(trials, fname="test_trials.jld2")
"test_trials.jld2"

julia> trials_reloaded = load("test_trials.jld2");

julia> trials_reloaded[collect(keys(trials_reloaded))[1]] == trials
true

julia> simulated_effects = SimulatedEffects();

julia> writeJLD2(simulated_effects, fname="test_simulated_effects.jld2")
"test_simulated_effects.jld2"

julia> simulated_effects_reloaded = load("test_simulated_effects.jld2");

julia> simulated_effects_reloaded[collect(keys(simulated_effects_reloaded))[1]] == simulated_effects
true

julia> trials, _simulated_effects = GBCore.simulatetrials(genomes = GBCore.simulategenomes(n=10, verbose=false), n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=10, verbose=false);

julia> tebv = analyse(trials, max_levels=50, verbose=false);

julia> writeJLD2(tebv, fname="test_tebv.jld2")
"test_tebv.jld2"

julia> tebv_reloaded = load("test_tebv.jld2");

julia> tebv_reloaded[collect(keys(tebv_reloaded))[1]] == tebv
true

GBIO.writedelimited — Method

writedelimited(genomes::Genomes, sep::String = "	", fname::Union{Missing,String} = missing)::String

Save Genomes struct as a string-delimited (default=) file. Each row corresponds to a locus-allele combination. The first 4 columns correspond to the chromosome, position, all alleles in the locus (delimited by |), and the specific allele. The subsequency columns refer to the samples, pools, entries or genotypes.

Notes:

Extension name should be '.tsv', '.csv', or '.txt'.

Header lines and comments are prefixed by '#'.

There are 2 header lines prefixed by '#', e.g.:

header line 1: "chrom,pos,allalleles,allele,entry1,entry_2"

header line 2: "chrom,pos,allalleles,allele,population1,population_1"

Examples

julia> genomes = GBCore.simulategenomes(n=2, verbose=false);

julia> writedelimited(genomes, fname="test_genomes.tsv")
"test_genomes.tsv"

GBIO.writedelimited — Method

writedelimited(phenomes::Phenomes, sep::String = "	", fname::Union{Missing,String} = missing)::String

Save Phenomes struct as a string-delimited (default=) file. Each row corresponds to a samples, pools, entries or genotypes. The first 2 columns correspond to the entry and population names. The subsequency columns refer to the traits containing the phenotype values of each entry. Note that the extension name should be '.tsv', '.csv', or '.txt'.

Notes:

Extension name should be '.tsv', '.csv', or '.txt'.

Header line and comments are prefixed by '#'.

There is 1 header line prefixed by '#', e.g.: "entry,population,trait1,trait2,trait_3"

Examples

julia> phenomes = Phenomes(n=2, t=2); phenomes.entries = ["entry_1", "entry_2"]; phenomes.traits = ["trait_1", "trait_2"];

julia> writedelimited(phenomes, fname="test_phenomes.tsv")
"test_phenomes.tsv"

GBIO.writedelimited — Method

writedelimited(trials::Trials, sep::String = "	", fname::Union{Missing,String} = missing)::String

Save Trials struct as a string-delimited (default=) file. Each row corresponds to a samples, pools, entries or genotypes. The first 10 columns correspond to:

years

seasons

harvests

sites

entries

populations

replications

blocks

rows

cols

The subsequency columns refer to the traits containing the phenotype values.

Notes:

Extension name should be '.tsv', '.csv', or '.txt'.

Header line and comments are prefixed by '#'.

There is 1 header line prefixed by '#', e.g.: "years,seasons,harvests, ..., trait1,tratit2,trait_3"

Examples

julia> trials = Trials(n=1, t=2); trials.years = ["year_1"]; trials.seasons = ["season_1"]; trials.harvests = ["harvest_1"]; trials.sites = ["site_1"]; trials.entries = ["entry_1"]; trials.populations = ["population_1"]; trials.replications = ["replication_1"]; trials.blocks = ["block_1"]; trials.rows = ["row_1"]; trials.cols = ["col_1"]; trials.traits = ["trait_1", "trait_2"];

julia> writedelimited(trials, fname="test_trials.tsv")
"test_trials.tsv"

GBIO.writevcf — Method

writevcf(genomes::Genomes; fname::Union{Missing,String}, ploidy::Int64=0, max_depth::Int64=100, n_decimal_places::Int64=4)::String

Save Genomes struct as a variant call format (VCF version 4.2) file.

Examples

julia> genomes_1 = GBCore.simulategenomes(n=2, verbose=false);

julia> writevcf(genomes_1, fname="test_genomes_1.vcf")
"test_genomes_1.vcf"

julia> genomes_2 = GBCore.simulategenomes(n=2, n_alleles=3, verbose=false);

julia> genomes_2.allele_frequencies = round.(genomes_2.allele_frequencies .* 4) ./ 4;

julia> writevcf(genomes_2, fname="test_genomes_2.vcf", ploidy=4)
"test_genomes_2.vcf"